Fix failure to compile debug extension and client with dart2js
